requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
frozendict==2.4.6
html5lib==1.1
idna==3.10
lxml==5.3.0
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.7
platformdirs==4.3.6
psycopg==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.48
